Update giraffe sprite path in App.cpp
Changed the path to the giraffe sprite in the App.cpp file. The previous path led to an outdated image file (grid-0003.png). It is now updated to point to the new giraffe sprite located in the assets/sprites directory. This change will help maintain better organization of the asset files and ensure that the app displays the most recent and relevant graphics.